[FEATURE] Add `failOnMatch` option
This option, when `"true"` will fail the checker when the given pattern
matches the commit message. This allows for a simpler kind check for
negative conditions such as "The commit message does not start with
'fixup'."

I rewrote some test cases here so that they were grouped into input
types. This allowed me to use the same inputs for both the default
settings and also the `failOnMatch` setting. Each of the "default" tests
are exactly the same as they were before, with the "via `failOnMatch`"
tests being the tests that were actually added.